SemanticSerachTesting/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
biopython>=1.84
transformers>=4.42.0
torch>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.28.0
streamlit>=1.33.0
